+++ tests/test_00_dxf_low_level_structs/test_020_layer_name_validator.py
@@ -1,6 +1,6 @@
 # Copyright (c) 2019 Manfred Moitzi
 # License: MIT License
-from ezdxf.lldxf.validator import is_valid_layer_name, is_adsk_special_layer
+from ezdxf.lldxf.validator import is_valid_layer_name
 
 
 def test_invalid_layer_name():
@@ -20,6 +20,6 @@ def test_invalid_layer_name():
 
 
 def test_is_adsk_special_layer():
-    assert is_adsk_special_layer('*adsk_xyz') is True
-    assert is_adsk_special_layer('*ADSK_xyz') is True
-    assert is_adsk_special_layer('ADSK_xyz*') is False
+    assert is_valid_layer_name('*adsk_xyz') is True
+    assert is_valid_layer_name('*ADSK_xyz') is True
+    assert is_valid_layer_name('ADSK_xyz*') is False